Refactor: Introduce public `:base:api`, move stuff into it.
Now there are two modules: `:base:api` and `:base:impl`.
