+++ api/dynamic/build.gradle.kts
@@ -6,7 +6,7 @@ dependencies {
     api(project(":api:public"))
     api(project(":rt:support"))
 
-    implementation(project(":base"))
+    implementation(project(":base:impl"))
     implementation(project(":rt:engine"))
 }
