+++ base/api/src/main/kotlin/Extensible.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022 Yandex LLC
+ * Copyright 2023 Yandex LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.yandex.yatagan.core.graph
+package com.yandex.yatagan.base.api
 
 /**
  * An extension API, which allows an object to have associated data, accessible by strictly-typed keys.
