+++ base/impl/build.gradle.kts
@@ -1,0 +1,7 @@
+plugins {
+    id("yatagan.artifact")
+}
+
+dependencies {
+    api(project(":base:api"))
+}
