+++ codegen/impl/build.gradle.kts
@@ -5,7 +5,7 @@ plugins {
 dependencies {
     implementation(project(":api:public"))
     implementation(project(":core:graph:api"))
-    implementation(project(":base"))
+    implementation(project(":base:impl"))
     implementation(project(":codegen:poetry"))
     implementation(project(":lang:compiled"))
 }
