+++ codegen/impl/src/main/kotlin/ConditionGenerator.kt
@@ -18,12 +18,12 @@ package com.yandex.yatagan.codegen.impl
 
 import com.squareup.javapoet.ClassName
 import com.squareup.javapoet.TypeName
+import com.yandex.yatagan.base.api.Extensible
 import com.yandex.yatagan.codegen.poetry.ExpressionBuilder
 import com.yandex.yatagan.codegen.poetry.TypeSpecBuilder
 import com.yandex.yatagan.codegen.poetry.buildExpression
 import com.yandex.yatagan.core.graph.BindingGraph
 import com.yandex.yatagan.core.graph.BindingGraph.LiteralUsage
-import com.yandex.yatagan.core.graph.Extensible
 import com.yandex.yatagan.core.graph.normalized
 import com.yandex.yatagan.core.model.ConditionModel
 import com.yandex.yatagan.core.model.ConditionScope
