+++ core/graph/impl/build.gradle.kts
@@ -1,10 +1,10 @@
 plugins {
-    id("yatagan.artifact")
+    id("yatagan.implementation-artifact")
 }
 
 dependencies {
     api(project(":core:graph:api"))
 
-    implementation(project(":base"))
+    implementation(project(":base:impl"))
     implementation(project(":validation:format"))
 }
