+++ core/model/impl/build.gradle.kts
@@ -6,5 +6,5 @@ dependencies {
     api(project(":core:model:api"))
 
     implementation(project(":validation:format"))
-    implementation(project(":base"))
+    implementation(project(":base:impl"))
 }
