+++ lang/rt/build.gradle.kts
@@ -6,5 +6,5 @@ dependencies {
     api(project(":lang:common"))
 
     implementation(project(":api:public"))
-    implementation(project(":base"))
+    implementation(project(":base:impl"))
 }
