+++ validation/impl/build.gradle.kts
@@ -6,6 +6,6 @@ dependencies {
     api(project(":validation:api"))
     api(project(":validation:spi"))
 
-    implementation(project(":base"))
+    implementation(project(":base:impl"))
     implementation(project(":validation:format"))
 }
